Issue #1: Pipeline: Use explicit version to publish docs
The pages job didn't work before and missed the rsync command. The rsync
part was moved to the build jobs that write the content intended for
public to an out/ directory. The pages job simply renames that
directory.

The build artifacts are published to an explicit version of the
documentation.

During the documentation build, the version is read from the variable in
the CI/CD pipeline.