docs(keycloak-migration): Fix document structure
The entry file contained headings with subsections. All subsequent
content was structured under the second section from the introduction.
This looks strange, especially in the PDF artifact.

This changes fixes it by introducing a different entry file for Sphinx.
This file only contains to document title and the table of contents. It
introduces the section "Introduction". All other sections are on the
same level in the table of contents.