docs(keycloak-app): Reload configuration after web server customization
To activate a custom web server configuration for Keycloak,
the user needs to activate it.
The commands were missing.
I use the reload command to keep the downtime as low as possible.
And I advice the reader to first validate the web server configuration
before reloading the web server.

Issue #216