fix(univention-keycloak): check if mapper exists that maps uid
After git:ff1202fe4fc85c4e3cfae198eba42e92bfcab99c a new protocol mapper
was being added to the SAML client. Before it added itself it checked if
an existing mapper with the same name exists. However it did not check
if there was an existing mapper that already mapped the uid or the URN.
Since a UID mapper is already automatically created due to the SAML
metadata of the UMC there were 2 UID mappers. This lead to the UID
appearing twice in the SAML assertion. This double UID was just
concatenated by the UMC, so the username "anna" became "annaanna", which
caused login to fail.

Check if there already is ANY mapper for the client that maps uid or the
uid URN.

Issue univention/dev-issues/dev-incidents#44
Bug #57420