Bug #56474: Fix broken permanent links in changelog
Sphinx creates permanent links from the headings in a file.
For headings not starting with a letter, it takes the part starting with
a letter, such as #ucs-1.
And other headings receive a numbered heading, such as #id1.

The changelog has multiple versions with the `ucs-1` and `ucs-2` suffix
and versions without any letter. This results in "dynamic" links that
aren't permanent for the document.

To mitigate this problem, I added the prefix "Version" to every heading.
Now the heading links are truly permanent links.